+++ test/test_punctuation.py
@@ -21,8 +21,8 @@ from phonemizer.punctuation import Punctuation
 from phonemizer.phonemize import phonemize
 
 
-# True if we are using espeak>=1.50
-ESPEAK_150 = (EspeakBackend.version(as_tuple=True) >= (1, 50))
+# True if we are using espeak>=1.49.3
+ESPEAK_143 = (EspeakBackend.version(as_tuple=True) >= (1, 49, 3))
 
 
 @pytest.mark.parametrize(
@@ -54,8 +54,8 @@ def test_preserve(inp):
 @pytest.mark.parametrize(
     'text, output', [
         (['hi; ho,"'], ['haɪ ; hoʊ ,']),
-        (['hi; "ho,'], ['haɪ ; hoʊ ,'] if ESPEAK_150 else ['haɪ ;  hoʊ ,']),
-        (['"hi; ho,'], ['haɪ ; hoʊ ,'] if ESPEAK_150 else [' haɪ ; hoʊ ,'])])
+        (['hi; "ho,'], ['haɪ ; hoʊ ,'] if ESPEAK_143 else ['haɪ ;  hoʊ ,']),
+        (['"hi; ho,'], ['haɪ ; hoʊ ,'] if ESPEAK_143 else [' haɪ ; hoʊ ,'])])
 def test_preserve_2(text, output):
     marks = ".!;:,?"
     p = Punctuation(marks=marks)